+++ src/program/main.cpp
@@ -40,10 +40,6 @@
 #include <fcntl.h>
 #include <getopt.h>
 
-
-#define SOCKET_FILENAME "/tmp/libTAS.socket"
-
-// std::vector<std::string> shared_libs;
 Context context;
 
 static void print_usage(void)